install dependencies from Module::Build
cpm only supports cpanfile, but not reading a Build.PL script like cpanm
does, so we need to run it and use the build-in installation. Leon T
tells me this is not the best way of doing it, but he concurs it's the
only way unless cpm supports this format, or we use cpanm.