Use := properly
We could have changed all the `:=` to `:-` here as that means "use value
if there is none already".

`:=` is a desctructive operation that does the same thing to the
variable in place. (think like `s/.../.../` in perl)

`:=` is a side effect operation so we do it in (ignored) params of the
`:` no-op command.

```
: "${MAYBE:=something}"
```