Some basic Bash refactoring
Now we can refactor some of the larger scripts.

I always define a `main() ( ... )` function.

Use parens around the bodies to prevent side-effects.
Use curlies to keep side-effects (like initializing vars).

Call `main` at the end with one of:
```
main "$@"

[[ ${BASH_SOURCE[0]} != "$0" ]] || main "$@"
```

Use the second one if the file is expected to be source'd for some
reason.